requirements.txt
fitz==0.0.1.dev2
layoutparser==0.3.4
opencv-python==4.9.0.80
Pillow==10.3.0
requests==2.31.0
numpy==1.26.4